changed saving user avatar for WebSpeak v.2

+++ WebSpeak v.2/VitaliiMahlona/LearningLanguagesAngular/Controllers/AccountController.cs
@@ -298,6 +298,7 @@ namespace LearningLanguagesAngular.Controllers
                 Users currrentUser = await _userManager.GetUserAsync(User);
                 usersInfo.CurrentUser = currrentUser;
                 usersInfo.IsSignedIn = true;
+                usersInfo.Avatar = currrentUser.Avatar;
             }
             else
             {
@@ -338,7 +339,7 @@ namespace LearningLanguagesAngular.Controllers
         }
 
         [HttpPost("Account/Manage/PersonalInfo")]
-        public async Task<PersonalInfoViewModel> PersonalInfo([FromBody]PersonalInfoViewModel model)
+        public async Task<PersonalInfoViewModel> PersonalInfo([FromForm]PersonalInfoViewModel model)
         {
             PersonalInfoViewModel error = new PersonalInfoViewModel
             {
@@ -390,9 +391,12 @@ namespace LearningLanguagesAngular.Controllers
                 {
                     var currentUser = await _userManager.GetUserAsync(User);
 
-                    var fileOldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\UserImages", currentUser.Avatar);
+                    if (currentUser.Avatar != null)
+                    {
+                        var fileOldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\UserImages", currentUser.Avatar);
 
-                    System.IO.File.Delete(fileOldPath);
+                        System.IO.File.Delete(fileOldPath);
+                    }
 
                     var fileName = Path.GetFileName(model.Avatar.FileName);
